refactor(proxy): share cancellation deferral helper
